pyproject.toml
[project]
name = "dropzone47"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-telegram-bot>=22.3",
  "yt-dlp>=2025.8.27",
  "python-dotenv>=1.0.1",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
  "E", # pycodestyle
  "F", # pyflakes
  "I", # isort
  "B", # bugbear
]
ignore = [
  "E203", # whitespace before ':' (compat with black)
]

[tool.ruff.format]
quote-style = "preserve"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_untyped_defs = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
ignore_missing_imports = true
no_implicit_optional = true
strict_equality = true
exclude = [".venv"]
